backend/requirements.txt
# --- Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.31.0

# --- Database (The 2026 Standard) ---
prisma>=0.15.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0

# --- Security & Auth ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.9

# --- Utilities ---
httpx>=0.28.0
loguru>=0.7.0
